more floats

+++ src/Elite.Engine/Types/player_ship.cs
@@ -2,9 +2,9 @@
 {
     internal class player_ship
     {
-        internal int max_speed;
-        internal int max_roll;
-        internal int max_climb;
+        internal float max_speed;
+        internal float max_roll;
+        internal float max_climb;
         internal float max_fuel;
         internal float altitude;
         internal float cabtemp;
